+++ include/parser/commands.h
@@ -4,7 +4,7 @@
 #define COMMANDS_H
 
 #define NCON 3
-#define NCOM 28
+#define NCOM 29
 
 extern const int ncmd[2];
 extern const char *commands[NCON][NCOM];